refactor(localization): move LocalizationOverride + definitions to base module (ADR-020)
Mirrors the Authorization move: LocalizationOverride entity, its
QueryDefinition and ExportDefinition now live in Granit.Localization
so they can be consumed independently of EF Core.
