+++ src/Granit.Localization.EntityFrameworkCore/Internal/LocalizationOverrideConfiguration.cs
@@ -1,4 +1,4 @@
-using Granit.Localization.EntityFrameworkCore.Entities;
+using Granit.Localization.Domain;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
